requirements.txt
numpy
tensorflow
Pillow
Flask